P03_FootballBetting update 100/100

+++ 05. Entity Relations/Exercises/P03_FootballBetting/P03_FootballBetting.Data.Models/Game.cs
@@ -16,9 +16,9 @@
 
         public DateTime DateTime { get; set; }
 
-        public byte HomeTeamGoals { get; set; }
+        public int HomeTeamGoals { get; set; }
 
-        public byte AwayTeamGoals { get; set; }
+        public int AwayTeamGoals { get; set; }
 
         public double HomeTeamBetRate { get; set; }
